monitor: immediate handling of the worker count
The monitor grows a bit more responsible than it was.
It now directly handles the workers.

It only waits for those just spawned workers in `ensure_workers`.

The previous too imprecise/racy assertion yielded occasional
failures because of real-life variation of all the monitored
workers count (not easy to have under test).